+++ src/main.rs
@@ -1,8 +1,7 @@
-use std::sync::Arc;
-
 use structopt::StructOpt;
 
 use httpmock::standalone::start_standalone_server;
+use std::path::PathBuf;
 
 /// Holds command line parameters provided by the user.
 #[derive(StructOpt, Debug)]
@@ -11,6 +10,10 @@ pub struct CommandLineParameters {
     pub port: u16,
     #[structopt(short, long)]
     pub expose: bool,
+    #[structopt(short, long)]
+    pub static_mock_dir: Option<PathBuf>,
+    #[structopt(short, long)]
+    pub disable_access_log: bool,
 }
 
 #[tokio::main]
@@ -19,11 +22,25 @@ async fn main() {
 
     let params: CommandLineParameters = CommandLineParameters::from_args();
 
+    log::info!("██╗  ██╗████████╗████████╗██████╗ ███╗   ███╗ ██████╗  ██████╗██╗  ██╗");
+    log::info!("██║  ██║╚══██╔══╝╚══██╔══╝██╔══██╗████╗ ████║██╔═══██╗██╔════╝██║ ██╔╝");
+    log::info!("███████║   ██║      ██║   ██████╔╝██╔████╔██║██║   ██║██║     █████╔╝");
+    log::info!("██╔══██║   ██║      ██║   ██╔═══╝ ██║╚██╔╝██║██║   ██║██║     ██╔═██╗");
+    log::info!("██║  ██║   ██║      ██║   ██║     ██║ ╚═╝ ██║╚██████╔╝╚██████╗██║  ██╗");
+    log::info!("╚═╝  ╚═╝   ╚═╝      ╚═╝   ╚═╝     ╚═╝     ╚═╝ ╚═════╝  ╚═════╝╚═╝  ╚═╝");
+
     log::info!(
         "Starting {} server V{}",
         env!("CARGO_PKG_NAME"),
         env!("CARGO_PKG_VERSION")
     );
 
-    start_standalone_server(params.port, params.expose).await;
+    start_standalone_server(
+        params.port,
+        params.expose,
+        params.static_mock_dir,
+        !params.disable_access_log,
+    )
+    .await
+    .expect("an error occurred during mock server execution");
 }