+++ tests/internal/large_body_test.rs
@@ -1,9 +1,6 @@
 extern crate httpmock;
 
-use std::io::Read;
-
-use isahc::{get, get_async, Body, Request, RequestExt};
-use regex::Replacer;
+use isahc::{Request, RequestExt};
 
 use httpmock::MockServer;